requirements.txt
Django==3.1.4
google-cloud-bigquery==2.4.0
